+++ iftttie/dataclasses_.py
@@ -2,9 +2,9 @@ from __future__ import annotations
 
 from dataclasses import dataclass, field
 from datetime import datetime
+from pickle import loads
 from sqlite3 import Row
 from typing import Any, Optional
-from ujson import loads
 
 from iftttie.enums import ValueKind